# BACK-END/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Noroff E-Commerce ",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/products/": {
      "get": {
        "tags": [
          "Product"
        ],
        "description": "Show all products",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Product"
        ],
        "description": "Create a Product by Admin",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Create a Product by Admin",
            "required": true,
            "schema": {
              "$ref": "#/definitions/createProduct"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/admin/": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Admin Page needs token. This route produces render adminPanel.ejs",
        "produces": [
          "text/html"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/brands": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Admin Get Brands Page",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/categories": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Admin Get Categories Page",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/allusers": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Admin Page needs token. This route gets all users by Admin. This route produces render adminPanel.ejs user routes for json response",
        "produces": [
          "text/html"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/orders": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Admin Page needs token. This route gets all users by Admin. This route produces render adminPanel.ejs user routes for json response",
        "produces": [
          "text/html"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/login": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "This is the Login Page for only admin",
        "produces": [
          "application/json"
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "description": "Login for Admin for front page",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Login for Admin for front page",
            "required": true,
            "schema": {
              "$ref": "#/definitions/frontadminlogin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "createProduct": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Swagger Product"
        },
        "description": {
          "type": "string",
          "example": "Swagger Description"
        },
        "price": {
          "type": "number",
          "example": 1000
        },
        "quantity": {
          "type": "number",
          "example": 10
        },
        "date_added": {
          "type": "string",
          "example": "2022-01-01"
        },
        "imgurl": {
          "type": "string",
          "example": "https://example.com/image.jpg"
        },
        "status": {
          "type": "string",
          "example": "available"
        },
        "brandName": {
          "type": "string",
          "example": "Swagger Brand"
        },
        "categoryName": {
          "type": "string",
          "example": "Swagger Category"
        },
        "discount": {
          "type": "number",
          "example": 4
        }
      },
      "required": [
        "name",
        "description",
        "price",
        "quantity",
        "date_added",
        "imgurl",
        "status",
        "brandName",
        "categoryName",
        "discount"
      ]
    },
...
    "frontadminlogin": {
      "type": "object",
      "properties": {
        "identifier": {
          "type": "string",
          "example": "Admin"
        },
        "password": {
          "type": "string",
          "example": "P@ssword2023"
        }
      },
      "required": [
        "identifier",
        "password"
      ]
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "in": "header",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}